Change activities names
RecipeAdapter2 -> UserRecipeAdapter

DisplayActivity -> DetailsActivity

DisplayActivity2 -> UserDetailsActivity
