+++ app/src/main/java/com/example/project_1_menu_maker/UserDetailsActivity.java
@@ -1,37 +1,30 @@
 package com.example.project_1_menu_maker;
 
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.room.Room;
-
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Parcel;
 import android.text.method.ScrollingMovementMethod;
 import android.util.Log;
-import android.view.View;
-import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.bumptech.glide.Glide;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.room.Room;
+
+import com.example.project_1_menu_maker.db.AppDatabase;
+import com.example.project_1_menu_maker.db.RecipeDAO;
 import com.example.project_1_menu_maker.db.Recipes;
 import com.example.project_1_menu_maker.db.UserDAO;
 import com.example.project_1_menu_maker.models.Recipe;
+import com.google.android.material.snackbar.Snackbar;
 import com.squareup.picasso.Picasso;
 
 import org.parceler.Parcels;
-import com.example.project_1_menu_maker.db.AppDatabase;
-import com.example.project_1_menu_maker.db.RecipeDAO;
-import com.example.project_1_menu_maker.models.Recipe;
-import com.google.android.material.snackbar.Snackbar;
-
-import org.json.JSONObject;
 
 import java.util.List;
 
-public class DisplayActivity extends AppCompatActivity {
+public class UserDetailsActivity extends AppCompatActivity {
     private static final String USER_ID_KEY = "com.example.project_1_menu_maker.db.userIdKey";
 //    private static final String PREFERENCES_KEY = "com.example.project_1_menu_maker.db.PREFERENCES_KEY";
 
@@ -55,7 +48,7 @@ public class DisplayActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_display);
+        setContentView(R.layout.activity_user_details);
 
         getDatabase();
 
@@ -69,7 +62,7 @@ public class DisplayActivity extends AppCompatActivity {
 
 
     private void wireUp(){
-        mRecipe = Parcels.unwrap(getIntent().getParcelableExtra("recipe"));
+        myRecipes = Parcels.unwrap(getIntent().getParcelableExtra("recipe"));
         mUserId = getIntent().getIntExtra(USER_ID_KEY, -1);
 
         tvMealTitle = findViewById(R.id.tvMealTitle);
@@ -115,14 +108,14 @@ public class DisplayActivity extends AppCompatActivity {
     }
 
     private void deleteRecipeFromUser(){
-        mRecipeDAO.deleteUserSpecificRecipeInDB(mUserId, mRecipe.getMealId());
+        mRecipeDAO.deleteUserSpecificRecipeInDB(mUserId, myRecipes.getMealId());
     }
 
     private boolean checkForRecipeInDB(){
         // use mRecipe and mUserId to check through DB
-        List<Recipes> myRecipes = mRecipeDAO.getAllUserRecipes(mUserId);
-        for(Recipes myRecipe : myRecipes){
-            if(myRecipe.getMealId() == mRecipe.getMealId()){
+        List<Recipes> myRecipeA = mRecipeDAO.getAllUserRecipes(mUserId);
+        for(Recipes myRecipe : myRecipeA){
+            if(myRecipe.getMealId() == myRecipes.getMealId()){
                 return true;
             }
         }
@@ -130,8 +123,8 @@ public class DisplayActivity extends AppCompatActivity {
     }
 
     private void loadRecipe(){
-        mRecipe = Parcels.unwrap(getIntent().getParcelableExtra("recipe"));
-        if(mRecipe == null){
+        myRecipes = Parcels.unwrap(getIntent().getParcelableExtra("recipe"));
+        if(myRecipes == null){
             Log.e("Big Fuckin error with", "DisplayActivity, Load Recipe (recipe is not in extra)");
             return;
         }
@@ -142,18 +135,14 @@ public class DisplayActivity extends AppCompatActivity {
             mFavoriteBtn.setImageResource(android.R.drawable.star_big_on);
         }
 
-        myRecipes = new Recipes(mUserId, mRecipe.getMealId(),
-                mRecipe.getTitle(), mRecipe.getCategory(),
-                mRecipe.getMealThumb(), mRecipe.getArea(),
-                mRecipe.getIngredients(), mRecipe.getInstruction());
 
-        tvMealTitle.setText(mRecipe.getTitle());
-        String ImageUrl = mRecipe.getMealThumb();
+        tvMealTitle.setText(myRecipes.getTitle());
+        String ImageUrl = myRecipes.getMealThumb();
 
         Log.d("Image", "bind: "+ ImageUrl);
         Picasso.get().load(ImageUrl).into(ivImage);
-        tvIngredient.setText("Ingredients:\n" + mRecipe.getIngredients());
-        tvInstruction.setText(mRecipe.getInstruction());
+        tvIngredient.setText("Ingredients:\n" + myRecipes.getIngredients());
+        tvInstruction.setText(myRecipes.getInstruction());
     }
 
     private void snackMaker(String message){
@@ -173,8 +162,8 @@ public class DisplayActivity extends AppCompatActivity {
                 .getUserDAO();
     }
 
-    public static Intent intentFactory(Context context, int userId, Recipe recipe){
-        Intent intent = new Intent(context, DisplayActivity.class);
+    public static Intent intentFactory(Context context, int userId, Recipes recipe){
+        Intent intent = new Intent(context, UserDetailsActivity.class);
         intent.putExtra(USER_ID_KEY, userId);
         intent.putExtra("recipe", Parcels.wrap(recipe));
         return intent;
