+++ app/src/main/res/layout/activity_user_details.xml
@@ -5,7 +5,7 @@
     android:id="@+id/layoutDisplayActivity"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DisplayActivity">
+    tools:context=".DetailsActivity">
 
     <ImageView
         android:id="@+id/ivImage"
